+++ crates/miden-lib/src/account/interface/mod.rs
@@ -6,7 +6,6 @@ use alloc::vec::Vec;
 use miden_objects::Word;
 use miden_objects::account::{Account, AccountCode, AccountId, AccountIdPrefix, AccountType};
 use miden_objects::assembly::mast::{MastForest, MastNode, MastNodeId};
-use miden_objects::crypto::dsa::rpo_falcon512;
 use miden_objects::note::{Note, NoteScript, PartialNote};
 use miden_objects::transaction::TransactionScript;
 use thiserror::Error;
@@ -16,6 +15,7 @@ use crate::account::components::{
     basic_fungible_faucet_library,
     basic_wallet_library,
     multisig_library,
+    no_auth_library,
     rpo_falcon_512_acl_library,
     rpo_falcon_512_library,
 };
@@ -147,9 +147,13 @@ impl AccountInterface {
                     component_proc_digests
                         .extend(rpo_falcon_512_acl_library().mast_forest().procedure_digests());
                 },
-                AccountComponentInterface::AuthRpoFalconMultisig(_) => {
+                AccountComponentInterface::AuthRpoFalcon512Multisig(_) => {
                     component_proc_digests
                         .extend(multisig_library().mast_forest().procedure_digests());
+                },
+                AccountComponentInterface::AuthNoAuth => {
+                    component_proc_digests
+                        .extend(no_auth_library().mast_forest().procedure_digests());
                 },
                 AccountComponentInterface::Custom(custom_procs) => {
                     component_proc_digests
@@ -267,25 +271,15 @@ impl From<&Account> for AccountInterface {
     fn from(account: &Account) -> Self {
         let components = AccountComponentInterface::from_procedures(account.code().procedures());
         let mut auth = Vec::new();
-        components.iter().for_each(|interface| {
-            match interface {
-                // RpoFalcon512, RpoFalcon512Acl, and RpoFalconMultisig use the same RpoFalcon512
-                // auth scheme
-                AccountComponentInterface::AuthRpoFalcon512(storage_index)
-                | AccountComponentInterface::AuthRpoFalcon512Acl(storage_index)
-                | AccountComponentInterface::AuthRpoFalconMultisig(storage_index) => {
-                    auth.push(AuthScheme::RpoFalcon512 {
-                        pub_key: rpo_falcon512::PublicKey::new(
-                            account
-                                .storage()
-                                .get_item(*storage_index)
-                                .expect("invalid storage index of the public key"),
-                        ),
-                    })
-                },
-                _ => {},
+
+        // Find the auth component and extract all auth schemes from it
+        // An account should have only one auth component
+        for component in components.iter() {
+            if component.is_auth_component() {
+                auth = component.get_auth_schemes(account.storage());
+                break;
             }
-        });
+        }
 
         Self {
             account_id: account.id(),